Updated Logic for UA based device detection.
    * Added fallback for detecting UA's in the abscence of Mongo backed DB's
