+++ src/Native5/Route/DeviceDetection/resources/.gitignore
